code cleanup, a litte refactoring and some code comments
